rename: Home 페이지 파일 위치 이동
