+++ src/home/pages/Home/Home.tsx
@@ -1,0 +1,19 @@
+import { DrawerRanking } from '../../components/DrawerRanking/DrawerRanking';
+import { Notification } from '../../components/Notification/Notification';
+import { SearchKeyword } from '../../components/SearchKeyword/SearchKeyword';
+import SocialNetworkService from '../../components/SocialNetworkService/SocialNetworkService';
+
+import { StyledComponentContainer } from './Home.style';
+
+export const Home = () => {
+  return (
+    <>
+      <Notification />
+      <StyledComponentContainer>
+        <DrawerRanking />
+        <SearchKeyword />
+        <SocialNetworkService />
+      </StyledComponentContainer>
+    </>
+  );
+};
